examples/python/requirements.txt
hiphops-hook
flask>=2.3.0
python-dotenv>=1.0.0